Refactor destroy, add update method in StudentService

+++ src/main/java/com/codecool/ccms/api/student/StudentService.java
@@ -1,6 +1,8 @@
 package com.codecool.ccms.api.student;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -29,7 +31,30 @@ public class StudentService {
        return studentRepository.findOne(id);
     }
 
-    public void destroy(String id) {
+    public ResponseEntity destroy(String id) {
+        Student student = getById(id);
+        if (student == null)
+            return new ResponseEntity(HttpStatus.NOT_FOUND);
         studentRepository.delete(id);
+        return new ResponseEntity(HttpStatus.OK);
+    }
+
+    public ResponseEntity update(String id, Student student) {
+
+        Student studentToUpdate = getById(id);
+        if (studentToUpdate == null) {
+            return null;
+        }
+        if (student.getFirstName() != null) {
+            studentToUpdate.setFirstName(student.getFirstName());
+        }
+        if (student.getLastName() != null) {
+            studentToUpdate.setLastName(student.getLastName());
+        }
+        if (student.getKlass() != null) {
+            studentToUpdate.setKlass(student.getKlass());
+        }
+        studentRepository.save(studentToUpdate);
+        return new ResponseEntity(getById(id), HttpStatus.OK);
     }
 }
